fixed sum of float column because it failed in release mode

+++ src/tightdb/table_accessors.hpp
@@ -1281,10 +1281,10 @@ public:
         return *Base::m_query;
     };
 
-    double sum(std::size_t* resultcount=NULL, std::size_t start=0,
+    float sum(std::size_t* resultcount=NULL, std::size_t start=0,
                std::size_t end = std::size_t(-1), std::size_t limit=std::size_t(-1)) const
     {
-        return Base::m_query->m_impl.sum<double, float>(col_idx, resultcount, start, end, limit);
+        return Base::m_query->m_impl.sum<float, float>(col_idx, resultcount, start, end, limit);
     }
 
     float maximum(std::size_t* resultcount=NULL, std::size_t start=0,
